feat(tailwind): publish tailwind preset
